Move rotor.cpp and rotor.h from  elast to core directory
