+++ src/core/rotor.cpp
